updating cod
